Added the discount and its views all over the place. It works but might blow up.

+++ Candyshop/Models/Candy.cs
@@ -14,7 +14,7 @@ namespace Candyshop.Models
         public string ImageUrl { get; set; }
         public string ImageThumbnailUrl { get; set; }
         public bool IsOnSale { get; set; }
-        public int SalePercentage { get; set; }
+        public decimal SalePercentage { get; set; }
         
         public DateTime SaleStartDate { get; set; }
         public DateTime SaleEndDate { get; set; }
